Location#population needs to be long (Asia has a population of 3812366000 :)
Former-commit-id: 3a7114d2ab31d92c3d3664707ece2e7bde9ed75c

+++ palladian-core/src/main/java/ws/palladian/extraction/location/GeonamesImporter.java
@@ -172,7 +172,7 @@ public final class GeonamesImporter {
         location.setLatitude(Double.valueOf(parts[4]));
         location.setPrimaryName(parts[1]);
         location.setAlternativeNames(alternateNames);
-        location.setPopulation(Integer.valueOf(parts[14]));
+        location.setPopulation(Long.valueOf(parts[14]));
         location.setType(mapType(parts[6]));
         return location;
     }
